Merge in PLZBBT/erlang-project from PLZBBT-122-back-end-maken to feature/PLZBBT-8-als-docent-wil-ik-de-gegevens-van-een-skilltree-kunnen-bijwerken-zodat-deze

* commit '96233b6f8317123eb059360043c622905dbcf3f0':
  backend werkend
  Changed "docent" to "gebruiker'

+++ Code/Java-REST-API/src/main/java/nl/han/oose/project/resources/dto/SkilltreeDTO.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class SkilltreeDTO {
-    private double id;
+    private int id;
     private String title;
     private String description;
     private List<NodeDTO> nodes = new ArrayList<>();
@@ -12,7 +12,7 @@ public class SkilltreeDTO {
 
     public SkilltreeDTO(){}
 
-    public SkilltreeDTO(double id, String title, String description) {
+    public SkilltreeDTO(int id, String title, String description) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -20,7 +20,7 @@ public class SkilltreeDTO {
         this.edges = edges;
     }
 
-    public void setId(double id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -40,7 +40,7 @@ public class SkilltreeDTO {
         this.edges = edges;
     }
 
-    public double getId() {
+    public int getId() {
         return id;
     }
